+++ rust/tests/integration/producer.rs
@@ -691,6 +691,121 @@ fn produce_no_mid_single_encoding_without_rid_or_ssrc() {
 }
 
 #[test]
+fn produce_single_encoding_no_temporal_is_simple() {
+    future::block_on(async move {
+        let (_worker, _router, transport_1, _transport_2) = init().await;
+
+        let producer = transport_1
+            .produce(ProducerOptions::new(
+                MediaKind::Video,
+                RtpParameters {
+                    mid: None,
+                    codecs: vec![RtpCodecParameters::Video {
+                        mime_type: MimeTypeVideo::Vp8,
+                        payload_type: 96,
+                        clock_rate: NonZeroU32::new(90000).unwrap(),
+                        parameters: RtpCodecParametersParameters::default(),
+                        rtcp_feedback: vec![RtcpFeedback::Nack, RtcpFeedback::NackPli],
+                    }],
+                    header_extensions: vec![],
+                    encodings: vec![RtpEncodingParameters {
+                        ssrc: Some(11111111),
+                        ..RtpEncodingParameters::default()
+                    }],
+                    rtcp: RtcpParameters {
+                        cname: Some("test".to_string()),
+                        ..RtcpParameters::default()
+                    },
+                    msid: None,
+                },
+            ))
+            .await
+            .expect("Failed to produce single-encoding no temporal video");
+
+        assert_eq!(producer.r#type(), ProducerType::Simple);
+    });
+}
+
+#[test]
+fn produce_single_encoding_temporal_is_svc() {
+    future::block_on(async move {
+        let (_worker, _router, transport_1, _transport_2) = init().await;
+
+        let producer = transport_1
+            .produce(ProducerOptions::new(
+                MediaKind::Video,
+                RtpParameters {
+                    mid: None,
+                    codecs: vec![RtpCodecParameters::Video {
+                        mime_type: MimeTypeVideo::Vp8,
+                        payload_type: 96,
+                        clock_rate: NonZeroU32::new(90000).unwrap(),
+                        parameters: RtpCodecParametersParameters::default(),
+                        rtcp_feedback: vec![RtcpFeedback::Nack, RtcpFeedback::NackPli],
+                    }],
+                    header_extensions: vec![],
+                    encodings: vec![RtpEncodingParameters {
+                        ssrc: Some(11111111),
+                        scalability_mode: "L1T3".parse().unwrap(),
+                        ..RtpEncodingParameters::default()
+                    }],
+                    rtcp: RtcpParameters {
+                        cname: Some("test".to_string()),
+                        ..RtcpParameters::default()
+                    },
+                    msid: None,
+                },
+            ))
+            .await
+            .expect("Failed to produce single-encoding temporal video");
+
+        assert_eq!(producer.r#type(), ProducerType::Svc);
+    });
+}
+
+#[test]
+fn produce_multiple_encodings_is_simulcast() {
+    future::block_on(async move {
+        let (_worker, _router, transport_1, _transport_2) = init().await;
+
+        let producer = transport_1
+            .produce(ProducerOptions::new(
+                MediaKind::Video,
+                RtpParameters {
+                    mid: None,
+                    codecs: vec![RtpCodecParameters::Video {
+                        mime_type: MimeTypeVideo::Vp8,
+                        payload_type: 96,
+                        clock_rate: NonZeroU32::new(90000).unwrap(),
+                        parameters: RtpCodecParametersParameters::default(),
+                        rtcp_feedback: vec![RtcpFeedback::Nack, RtcpFeedback::NackPli],
+                    }],
+                    header_extensions: vec![],
+                    encodings: vec![
+                        RtpEncodingParameters {
+                            ssrc: Some(11111111),
+                            ..RtpEncodingParameters::default()
+                        },
+                        RtpEncodingParameters {
+                            ssrc: Some(22222222),
+                            ..RtpEncodingParameters::default()
+                        },
+                    ],
+                    rtcp: RtcpParameters {
+                        cname: Some("test".to_string()),
+                        ..RtcpParameters::default()
+                    },
+                    msid: None,
+                },
+            ))
+            .await
+            .expect("Failed to produce multiple-encoding video");
+
+        assert_eq!(producer.r#type(), ProducerType::Simulcast);
+    });
+}
+
+#[test]
 fn dump_succeeds() {
     future::block_on(async move {
         let (_worker, _router, transport_1, transport_2) = init().await;